requirements.txt
python-telegram-bot>=20.7
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
pytz>=2024.1
